requirements.txt
asgiref==3.6.0
attrs==22.2.0
cfgv==3.3.1
distlib==0.3.6
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
filelock==3.9.0
identify==2.5.18
inflection==0.5.1
jsonschema==4.17.3
nodeenv==1.7.0
platformdirs==3.0.0
pre-commit==3.0.4
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
virtualenv==20.19.0
